sorting out edge cases and ran tests

+++ src/Models/InboundRequest.php
@@ -32,7 +32,7 @@ class InboundRequest extends Model
     protected $casts = [
         'headers' => 'array',
         'response_headers' => 'array',
-        'duration_ms' => 'integer',
+        'duration_ms' => 'float',
         'status_code' => 'integer',
         'user_id' => 'integer',
     ];
